refactored by adding 2 new diroctorys adding PersonCard.tsx with PersonCard.css adding Hooks with useQRCodeScanner.ts and backuping old app by renaming to AppOld.tsx
